# Faster source for goroutine-local use