--- a/source.go
+++ b/source.go
@@ -9,17 +9,24 @@ import (
 // CSRandSource is a math/rand.Source that wraps crypto/rand.Reader.
 type CSRandSource struct{}
 
-// Int63 implements the math/rand.Source interface.  Will panic if an
-// error occurs when reading the underlying Reader.
-func (r CSRandSource) Int63() int64 {
-	var i int64
-
-	err := binary.Read(rand.Reader, binary.LittleEndian, &i)
-	if err != nil {
+// readUint64Bytes returns 8-bytes, suitable for pulling an uint64 out of.
+func readUint64Bytes() []byte {
+	buf := make([]byte, 8)
+	n, err := rand.Reader.Read(buf)
+	if err != nil || n != 8 {
 		panic("could not read random data")
 	}
 
-	return (i & 0x7fffffffffffffff)
+	return buf
+}
+
+// Int63 implements the math/rand.Source interface.  Will panic if an
+// error occurs when reading the underlying Reader.
+func (r CSRandSource) Int63() int64 {
+	buf := readUint64Bytes()
+	i := int64(binary.LittleEndian.Uint64(buf))
+
+	return i & 0x7fffffffffffffff
 }
 
 // Seed implements the math/rand.Source interface.  Does nothing.
@@ -30,14 +37,8 @@ func (r CSRandSource) Seed(s int64) {
 // Uint64 implements the math/rand.Source64 interface.  Will panic if an
 // error occurs when reading the underlying Reader.
 func (r CSRandSource) Uint64() uint64 {
-	var i uint64
-
-	err := binary.Read(rand.Reader, binary.LittleEndian, &i)
-	if err != nil {
-		panic("could not read random data")
-	}
-
-	return i
+	buf := readUint64Bytes()
+	return binary.LittleEndian.Uint64(buf)
 }
 
 var _ mrand.Source = CSRandSource{}