This package is great!  But syscalls are pretty expensive.  This PR introduces a new `CSStatefulRandSource` to amortize the cost of calling `getrandom` -- the idea is to use it with `math/rand.Rand`s you manage with a `sync.Pool`, so that we only do a syscall once every 512 calls to `Uint63()`.

I added benchmarks so you can see the difference between sources (the new one is `BenchmarkLocalStatefulFloat64`):

```
$ go test -run=XXX -bench=. -benchtime=2s -cpu=1,2,4,8 | tee -a new.txt
goos: linux
goarch: arm64
pkg: github.com/andrew-d/csmrand
BenchmarkGlobalFloat64            	 8557785	       277.6 ns/op	  28.82 MB/s	       8 B/op	       1 allocs/op
BenchmarkGlobalFloat64-2          	10388292	       214.2 ns/op	  37.34 MB/s	       8 B/op	       1 allocs/op
BenchmarkGlobalFloat64-4          	 5223753	       467.3 ns/op	  17.12 MB/s	       8 B/op	       1 allocs/op
BenchmarkGlobalFloat64-8          	 4475575	       525.8 ns/op	  15.21 MB/s	       8 B/op	       1 allocs/op
BenchmarkMathGlobalFloat64        	178649461	        13.41 ns/op	 596.59 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathGlobalFloat64-2      	65284251	        38.68 ns/op	 206.80 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathGlobalFloat64-4      	25149420	        93.25 ns/op	  85.79 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathGlobalFloat64-8      	18248013	       132.3 ns/op	  60.45 MB/s	       0 B/op	       0 allocs/op
BenchmarkLocalFloat64             	 8610688	       276.2 ns/op	  28.96 MB/s	       8 B/op	       1 allocs/op
BenchmarkLocalFloat64-2           	10693802	       228.7 ns/op	  34.98 MB/s	       8 B/op	       1 allocs/op
BenchmarkLocalFloat64-4           	 5481765	       462.1 ns/op	  17.31 MB/s	       8 B/op	       1 allocs/op
BenchmarkLocalFloat64-8           	 4479069	       528.3 ns/op	  15.14 MB/s	       8 B/op	       1 allocs/op
BenchmarkLocalStatefulFloat64     	130450280	        18.39 ns/op	 434.93 MB/s	       0 B/op	       0 allocs/op
BenchmarkLocalStatefulFloat64-2   	100000000	        21.57 ns/op	 370.85 MB/s	       0 B/op	       0 allocs/op
BenchmarkLocalStatefulFloat64-4   	78684411	        32.20 ns/op	 248.45 MB/s	       0 B/op	       0 allocs/op
BenchmarkLocalStatefulFloat64-8   	61156548	        33.85 ns/op	 236.36 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathLocalFloat64         	994451737	         2.437 ns/op	3282.96 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathLocalFloat64-2       	1000000000	         1.247 ns/op	6412.95 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathLocalFloat64-4       	1000000000	         0.6393 ns/op	12513.60 MB/s	       0 B/op	       0 allocs/op
BenchmarkMathLocalFloat64-8       	1000000000	         0.3249 ns/op	24621.40 MB/s	       0 B/op	       0 allocs/op
PASS
```

So we can get performance that is _close_ to using the `math/rand.Float64()` exported function with a CSPRNG, but we don't really have a prayer of matching the non-CS `math/rand.Source` :)

I believe the scaling limits we see with `BenchmarkLocalStatefulFloat64-8` (the throughput being roughly the same as `BenchmarkLocalStatefulFloat64-4`) are locking in the kernel around the urandom source.  Hopefully it will just magically get faster when https://twitter.com/FiloSottile/status/1476698207413493766 lands in kernels I can test on!