@andrew-d 2022-04-27
@bpowers - Also, pushed [tag v0.2.0](https://github.com/andrew-d/csmrand/releases/tag/v0.2.0) along with a new README and GitHub Actions.